[API][Admin] Enable contract tests for Product Association Type resource
